Normalize use of X<>, L<>, and C<> somewhat
In particular, change some X<foo> to L<foo>.

Not all L<foo> are generating working links yet; this is something for
future commits to handle. When a term is used frequently in a region of
text, the first occurrance is usually linked (with L<foo>), and later
ones are marked as code (C<foo>) or emphasized (I<foo>) as appropriate.