Use find-definitions for Type files
This cuts down on a lot of redundancy and allows for the same procedure
to be used in places other than Type files.

=head2 lines in Type files will no longer be assumed to methods, and
this is intentional. Something that would previously be written:

    =head2 foo

        method foo (|bar) { }

is now written

    =head2 method foo

with or without the code line below it. Definitions for things that are
both subs and methods are still possible with, e.g.:

    =head2 routine foo

        method foo (|bar) { }
        sub    for ($baz, |bar) { }

Files in the lib/Type dir still must be updated for this change