Add summaries to type docs
I didn't get around to adding them for the Set-related and exception classes yet, but all other currently documented classes should be covered.
Also I didn't add them for roles for now, because I'm not yet sure what's the best way to phrase it for them... should the summary for Rational be "Rational number" or "represents a number as numerator and denumerator" or "role for rational numbers"?