Gloss some grammar, redo metamethod name conflicts
The example was obviously expecting .WHY to echo the class
name.  That does not in fact work.  I tried to give the .WHY a
Pod value, however, though the example is AFAICT correct,
it does not compile as is, and also, the proposed "WHY"() trick
does not in fact work anymore if it indeed ever did.