Rename compile-time "constants" -> compile-time variables
This is in response to #GH46.  In the ticket, it is recommended to remove
the quotes around the word "constants", since this causes a link to not
work.  Also, it is recommended that the word "variables" is a better choice.
Thus "variables" was chosen since the values aren't really constants (which
is the sense, the quotes were trying to convey before).