Improve subtitles for type pages
Besides misc tweaking and homogenizing of the subtitles, the subtitles for
roles are now worded similarly to those of classes (i.e. so it would fit "An
object which has this role composed into it represents a(n) [SUBTITLE HERE]"),
rather than starting their subtitles with "Common role for".

Remember that the main purpose for the subtitles is to provide helpful context
and hints to reader who are skimming the type index pages
(http://doc.perl6.org/type.html, http://doc.perl6.org/type-basic.html, etc.),
which is why it's important for the subtitles to be phrased consistently
(especially between related classes/roles) and to be free of fluff and
distractions.