Another attempt to fix initialization of @!categories (gh #227)
The problem is that when pre-building the attribute hash and then
creating the Perl6::Documentable object with .new(|%args), array
values are itemized by he hash, and thus the wrong thing happens
when the default constructor assigns them to the respective
array attributes of the class via single-argument rule.

This commit overrides the default constructor to remove the
itemization, which fixes the problem.

A better long-term solution might be to refactor htmlify.p6 to not
pre-build the attribute initialization lists as hashes.