Improve search when no results found
Changed "Nothing found" to "Not in index," to suggest the searched item
may be on the site, but not indexed.

Added a "try web search" link that googles for the searched term on
the doc site. Chose Google instead of DuckDuckGo, because some stuff
wasn't giving any results on DDG, but Google did find some pages.