Revert "Improve typography"
This reverts commit 36e44716cd4ab8f3f0d557fdfc0ea7f007a9c8ea.

Given that most people who commented on #652 disliked the change in one way or
another, I am reverting this commit.

This, however, does not mean that fonts do not have to be changed, or that
some of the changes in this commit are not good. As all of the changes are in
one rather big commit, it is a bit hard to cherry-pick what is actually needed,
especially given that there is no final consensus on each particular thing.

In other words, this revert does not solve #652, but it gives us a good start
for working on the issues. Further discussion in #652 is welcome.

This also reverts 540a3403c6787b0542864998efd7563aca5076e9. It is a followup
to the typography changes, which may or may not make sense when applied alone.
The commit message says “reduce line-height” but in reality it increases it
compared to the versions before “Improve typography” commit. It also conflicts
with the revert anyway.