Remove TOC hide/show feature
It's only used on narrow screens, since we now show TOC on the side
on wide screens. It's also is half-broken and has been for at least
several months and no one noticed, so I doubt the feature is in
high demand.

By removing it we can get rid of Cookies jquery plugin and the associated
notice about our site using cookies.

.oO( mmm.... coookies )