Revert flawed test
This test assumes Pod::To::BigPage is actually installed and located in a repository that the run ... htmlify.p6 is going to see. For instance: perl6 -I/path/to/Pod-To-BigPage/lib t/00-load.t will not work even though the dependency on Pod::To::BigPage is met. To use run like this would require passing all the $*REPOs (@INC) into the executed script (htmlify.p6)