Proc::Async example with ｢react｣
Tackle issue #1468.

IMO in this case it is reasonable to expect that people will be
copying the example and reducing it to their needs, so I tried to
include everything that may be needed.

The logic for kill signals is questionable, I don't know if there's
anything better we can show.

This is not mergeable because of RT #132016. Either wait for the
ticket to be resolved or get rid of ｢Supply.merge｣ and ｢Supply.zip｣.