Index everything on unicode_texas page
Resolves issue #1373.

This means that from now on some things may appear more than once in
the search. We'll see how annoying it is. If it is, we can start
removing symbols that are explained elsewhere (but note that a lot of
things are only mentioned on unicode_texas page, so don't revert the
whole commit).