Change headline so that it does not contain the word class and a type
This is a workaround for a problem in which htmlify.p6 can be tripped
up when it encounters what looks like a type heading in a file which
is not present in the type directory. This will lead to subtitles,
i.e. descriptions, going missing on https://docs.perl6.org/type.html.
See https://irclog.perlgeek.de/perl6/2017-09-24#i_15210015