Add a short introduction to subroutines and precedence dropping.
Should help removing the TODO tag on this section.
The subroutine are now briefly explained in the section, and the example
contains a reference to the hyper method call ">>".

A new head2 section about precedence dropping has been created
with a short example about the difference on using a "standard" method call
and a precedence drop one.