Make Label method examples clearer
Changed the example for the next method so that it's output differs from
that of the last example. With the change, the numbers 5 through 10 are
printed, which makes it more clear that the loop has iterated a total of
ten times but only reached the print statement during 6 of those.

Also, added the output to the example, as I believe that is clearer and
allows better visualization of what occurred.