Render Enums differently in type graphs
This should resolve #72. The ticket says to “render enum types
differently from classes”, but does not specify how exactly. This
commit makes enums green.

I am sure that there are better (and prettier) ways to do it, but the
graph generation code is already unnecessarily convoluted, so perhaps
we need a rewrite that will take care of all these tiny issues?