Conditional operator with the same explaination as in 5to6-perlop.
Not sure if we need to keep two sections to explain the very same
operator, or if we should link them together or remove one of them
(leaving 5to6-perlop one).
For a 'nutshell' it does make sense to keep at least a reference
for such common operator. Please consider I've kept the original example.