Improve &exit documentation.
Specify that exit status code has to be interpreted by calling process, and
that LEAVE is prevented when using exit.
Remark that &exit should be used to terminate abnormally only the main process,
while exceptions should be used for other cases (e.g., methods).

This does not remove the TODO item attached, but I believe
there's not a better place to document exit since it is already documented
with the usage of MAIN and the phasers.