We'd rather not advice «» to those who may not know enough
«» is basically a jump from the shell hell to exactly the same but
on perl 6 level. Those who want to do it may do it, but if we really
want to document this, then it needs to come with a warning and a
blurb on what this may lead to.

Also, in that particular example there are just two arguments so
using «» is not really justified.

Half-reverts 38ed8fbc7f1005db4cea683bb8dad3324da196ac.