Changed to a non-trivial example
Still the benefits are not so clear, because it's not in a
multithreaded environment or any other where you can really see the
advantages of computing without side effects. When `==>` is eventually
parallelized; I guess we could see an actual improvement.

Anyway, this example is more of the "macro expansion" type, and its
real effect is moving the computation of the data structure to compile
time. Would that be faster? I would need to measure it, actually... My
hunch is... probably. Anyway, it's a different use case, with no
input, and I hope it helps someone to understand this better. If you
think it does not, feel free to revert or modify.

Closes #1231, for tood, I guess...