Changed some examples to a case that feels a bit more real
And shows intent refs #1748. Also separates code in different
examples, leaving the text as text, so it addresses 4 and 5. It closes

Still it should also show an example where the method chaining does
not work by using some data structure that can be used as the last
argument to something, but does not have that something as a method.