Adds accessor overriding example
I'm using the same wording mentioned in the first comment by
@ronaldxs (get well soon :-) ) and including an example that overrides
mutator and accessor, not using `Proxy` but just using multi methods.

I hope it's OK, so this closes #136, phew, but obviously if anyone has
anything to object please feel free to re-open and make suggestions
and/or pull requests.