Implement assets generator
We're coming up to having 5 JS files now. 2 years ago, the argument
to leave them unminified was that someone would be able to debug
JS problems on live site. Doesn't seem practical to me any more.
No reason to punish thousands of docs readers, for years, with extra
bandwidth and HTTP requests, on an off chance we'll save this
mythical "debugger" 20 seconds of not having to clone the repo and
run the dev web app that caters unminified assets.