Improve the IO::Socket::INET.new method description.
In another piece of the documentation (method .accept) the 'server mode'
is referenced, but it seems it is not declared anywhere else. Therefore,
it is useful to explain in the new method that when the socket is 'listening'
the 'server mode' is activated (i.e., listen and server mode are the same).