m/ / and / / examples added
Added examples to help illustrate difference between m/ / and / /.  Zoffix provided the examples.  I provided the learning.  An example for rx/ / is needed.  As a newbie, this info is very helpful and rare.  It is not in Think Perl 6, ttbomk.  And definitely not in Perl6Intro.com.  Needed info or something better.