Ban “RHS” and “LHS” abbreviations
By removing them from words.pws and eliminating a few usages from the
docs.

We're sometimes criticized for using lingo that is unfamiliar to
beginners or those coming from other languages. While RHS and LHS are
fairly understandable abbreviations, I think we can just use
“right-hand side” or similar in most of the cases without sacrificing
anything.

I'm not entirely sure what would be the best way to hyphenate
it (right-hand side or right hand side? But not to be confused with
right-hand-side, right?). It would be nice if we were consistent about
it within our docs.