Tests for tests spun off and included
Now it checks not only valid Pod6, but also valid syntax.

This is actually a workaround for #1951. It was a Rakudo error having
to do with the setting up of a compunit, which is created when the
file is interpreted. It happened, when it did, by the end of the
tests. It became less frequent when the tests were split in two
batches, but it still happens from time to time. Spinning off these
tests should solve it. Besides, testing test files for valid POD6 is
not so critical, so it might be a good idea to move this new one to an
author test.

Anyway, closes #1951. If it keeps happening, we'll try to find another
workaround.