Adding some documentation
Reviewing if the merging indicated in #1822 is possible, but well, it
is not, because they target a different set of files. It might still
be interesting to unify all these tests, and create a set of
extensions as well as regexes or tests they should meet. But not for
the time being.
So it closes #1822.