Clarifying list and item assignment.
This would close #1810. I have opted for deleting a single example and
clarifying it by explaining it together with another example, and
clarifying also what assignments we refer to in every line. It's a
little less dense-ish, but my impression is that if there were better
documentation supporting list and item context, this could probably go
easily by just saying that list assignment occurs in list context,
item assignment in item context. That documentation is not there yet,
but when it's there, we might want to come back here to rewrite it
completely.