Changes code to avoid false positive
In this case, it was counting the number of `http`, and it was getting
confused due to the double https in that URL. I have changed it to
count them only if they follow the `|` in an `L` sentence. Still
hacky, but no fake positives now. Thanks @coke for running this,
closes #2039.