Adds introduction to the Cool section of traps
Along the lines suggested by @AlexDaniel. After 29 comments and half a
dozen of commits, it's difficult to say if this is solved or not, but
I think that it addresses @alexdaniel's (by copying it almost
literally) and thus closes #992.

I would suggest not to open this again. If there's a trap related to
Cool and Lists or Strings, please open a different issue and reference
this one, #992, for additional information.