Reviews relationship between ~~ and ACCEPTS.
Right now there is only a short definition in the first, and a pointer
to the last. It's nice to have the docs for ~~, since it's a single
point of entry; `ACCEPTS` page is generated from many pages. So I
would say it's better to close #1913, since it's not indexed and looks
like a sufficiently good reference.