Adds documentation for nativesize
Only in the way of explanation of the difference between native types
and native types with native size; added warning that it's not
spec. But this would close #2051.

I am going to close #1512 too. It does document all native types, and
clarifies the difference between them. This would be the last document
in #114, which was already closed, so I can't close it, but this would
basically be it. As the rest of the documents I have created, it's not
perfect, but it's there. It can be incrementally improved from now on.