Rewrite rakudo-specific note
Normally implementation-specific things are not documented, but in
this particular case we can prevent a lot of banging against a wall by
mentioning some known issues. In this case the behavior has changed
quite a bit so IMO it's a worthy mention.

See issue #1486 for more info on docs vs implementation-specific notes.