Adding example of stream (not dynamic stream variable) inheritance
I've adapted @bdfoy's example, included a bit of @jnthn explanation,
and kinda managed to make some sense. This would close #1226 if
everyone is satisfied with it.

This might be a trap instead of a part of the documentation. If
needed, a trap can be added and linked to this explanation.