Adding examples to resume
The example is in the spirit of @jnthn answer to @bdfoy in the
[StackOverflow
question](http://stackoverflow.com/q/43289842/2766176). However, some
things should be noted

* `.resume` is not peculiar to try statements, but can be used within
  any catch block.
* It might be a bit clearer if `.resume` returns to a statement that
  is not directly inside the CATCH block, to clarify what it does.

This would close #1274. As usual, if anyone is not satisfied with how
the OP is addressed, please reopen and, if possible, state clearly
what parts of the new examples and/or text are not satisfactory.