Introduce index page for 5->6 translation
There was no overview page for Perl 5 to Perl 6 translation.  Although
the nutshell page came close, I think it grew too large to be used as an
entry point, running the risk that people would miss the other pages that
deal with Perl 5 -> Perl 6 translation.

This commit introduces 5to-6-overview, which is basically a link to the
other pages.  The comments section was moved from the nutshell.