Eliminates intro
Which was probably OK for the book, but the reference to Perl is probably not needed. Since this intro is pretty much complete and adapted to the site, this closes #2145.

It might be interesting to add links to other parts of the documentation, but so far I think it's ready.