Unbust doc build
Argh. My guess is that the problem is in Pod::To::BigPage as it
expects a non-empty subtitle and starts crying if it's empty. That
should be fixed properly, but for now I just want to let the docs
website rebuild. Also, we should ot have these categories as table
rows, but I believe that the work for that is in progress.