Several improvements on this page
It actually started with #1566, trying to improve the description of
constants. However, while I was at it, I noticed references to
Synopsis, so I changed a few of them working on #2225.
This also fixes #2226 and, in fact, has generated rakudo/rakudo#2141
for a command line flag that was referenced here and was actually
working on Perl 6 (undocumented). Of course, there was a bit of #2223
and a bit of reflow.

I guess now it's back to work on #1566 or maybe #2225.