Deletes last reference to Perl 5
I have actually left the reference when talking about `use v6`. In
fact, it happens to me from time to time, and there's no way to
explain what it does other than referencing this.

I agree with @stmuk that it shouldn't be moved. It serves its purpose,
which is to serve as an introduction *for anyone*, not for Perl 5
users now that all those references have been eliminated. This closes #2191