Make pod-tables scroll horizontally on smaller views
If a table is too wide for a page a horizontal scroll bar will be added to the bottom.

If the page is focused, the right key will scroll right and left to come back.

I'm presuming this will work on mobile with finger dragging, but no testing set up for that at the moment.