Fixes link to Roles
This closes #2283. This link is probably better, since it's the declarator. The problem is that it maybe is the symptom for a problem, namely, /syntax/role generated at one particular point in time and no longer there for some fluke of the page generation. So thanks a lot for the report, and I'll investigate if that's the case.