Give minimal info about methods for Positional/Associative roles
We should at least mention the methods that *must* be implemented.
And we should refer to the extended information: since I had no clue
as to how to properly refer to this internal information in a link,
I've just use the URL (which apperently happens more often).  If this
is incorrect, please fix  :-)