Re-indexing of new as new (method)
I couldn't think of a better way to address #2309. Generated pages
don't have an intro, and adding one by hacking the generator would be
a slippery slope. Order can probably be changed, but once again, this
is not going to satisfy everyone; if it's hacked, same problem as
above. If we want to define an order for generated routine pages, this
would need metadata all across the site. Adding a prologue for
routines would be a nice thing, but once again, it would need changes
in the generator *and* adding many new and different pages.

So I've opted for just adding a new index entry, which does not
overwrite the previous one, but adds a new entry point for those who
want to know about new as a method. The section in the object language
page is good enough.

I'll close #2309, if you want to suggest something else, please feel
free to reopen with that suggestion.