Creates an initial version of the page
This goes towards #2330, but does not close it, since there are some
functions missing. It was created from the current `/language/testing`
text, and the text *is still there*. Will have to kinda redirect, or
otherwise sew the seam between the two pages that exist now, as well
as work with #2328.