BUILDALL has basically become an implementation detail
So remove it from the documentation.  If you really want to have your
own BUILDALL, you should *really*, **really** know what you're doing (and
know what you'll be throwing away in terms of optimizations).

So BUILDALL should be more in a "guts" type of documentation, if any.

Please revert if consensus is that BUILDALL should stay documented and
recommended as a way of unfluencing object creation.