Adds example to STORE and fixes indexing
This closes #2462. In the first case, a real example is added which
illustrates how to use. Indexing problem were due to the fact that it
was a `=head2` called `method` which is automatically considered for
indexing. That was also a problem, since it created a new section,
which it should not.