Revised and updated Dockerfile.
It's now in the Docker hub, readily available, and it's also open for
the public. Actually, the indicated error was due to the fact that the
original Dockerfile had to be run in the docs directory. It was more
oriented to testing and running locally than to actually contain. That
one has been moved to the no-docs tag, and is still available.

Anyway, this adds docs also, top of the README, and it closes #2525