Adds target destination and some reflow
Added back /syntax/role, which disappeared due to obscure indexing
rules (see #2575). Either when indexing for `does` was
added (apparently, adding an index term behind or before a header
suppresses the creation of a new page), or simply when the
X<|declarator,role> was changed, it disappeared. It's added back, refs #2568