update pod.pod6 re P<>. remove bad links
If `L<>` is used for an non-existent link (as an example), then the rendered pod will have a link error. In Placements, non-existent links are used. However, a link can be 'rendered' with `C<>` or `V<>`. Note that if a renderer cannot reach a link, then putting it as a target in a link `L<>` will not be much use as the user clicking on such a link will also not reach it.