Changes rewrite rule to avoid encoding
This goes towards fixing #935. I'll reopen it if it does not, I can't test it
locally.

Anyway, this is a hack that should probably be avoided at all
costs. If we can't provide a meaningful page, we should simply
eliminate it from the graph. I leave this issue open until that is
fixed.