Make .Str vs .IO example more explicit
As mentioned on #perl6 today, the original introductory text here
could lead one to think the difference was between `my IO $path` and
`my Str $path`, when it's actually in using `~` for getting the value
rather than `.absolute` or `.relative`.

This version uses `.Str` explicitly so it lines up more obviously with
the text.