Adds some more examples
Actually sigilless variables are all over the document, apparently
with the intent indicated before: to make it look like Python. I have
added a note, and also more examples that work the same, which I guess
illustrates also the More than One Way To Do It (theoretically
opposite the philosophy of Python). Closes #2699