Adds ParametricRoleGroupHOW
This part becomes a bit more complete, and allows us to understand the
inner workings (in fact, the *high-order-work* of Perl 6). Although
they are not (so much) intended for final user, they have been used,
for instance, in https://github.com/vrurg/Perl6-OO-Plugin

This closes #1306. Although there might be a few other metamodel
classes that have not been documented, they are probably deeper down
in the implementation and will never be seen by the user. There's no
request to use them, anyway.