Adds "constant" as syntax category
Also closes #2722. This is totally related to #1823 as expected, and
maybe needs a bit of explanation. The problem was not that it was
"special cased", but exactly the opposite: there was no special case
for constants. So a new "constant" category was added, and it was
added to the "syntax"-generating case so that the URL keeps being the
same. It's probably better to generate "syntax" URLs this way than
rely on some undocumented heuristics, as it did so far.