Suppressed reference to (non-existent) $?ENC
Which does not seem to have actually existed, or been needed. Maybe in
Synopsis, together with $?TAINT; this document makes a reference to
it: https://github.com/perl6/specs/blob/master/S29-functions.pod

However, it's not implemented (nor anything like that), so this
closes #2724 and moves around the surrounding to say the same,
although this needs to be restructured too.