Allow examples-compilation.t to pass for subs with multiline signatures
examples-compilation.t wraps code to test into a template and
heuristically adds some other bits, so that a bare

  method foo ($x)

would compile successfully. This includes adding empty blocks at
the end of a line when a situation like above is detected. Previously,
the heuristic was triggered also with this way to split a signature:

  submethod BUILD (
      ...
  ) {}

introducing a syntax error that wasn't the example code's fault.

Fixes #2750