Moved int to a new directory.
Since the files with which we are going to have trouble are int, num,
str and so on (if we decide to create them eventually), I have created
a new Native directory, then modified htmlify.p6 (yes, another hack)
so that it switches "native" for "type" kind. Tested locally, it
works.

This fixes #2749, fixes #935

An alternative would be to use .gitattributes combined with filters,
but that would need local modification of the downloaded
repos. Apparently, most repos implement the no-case-sensitive policy
in checked in files, so we'll go for that.