Add infix C«(^)», infix C«⊖» to operators
For now.  I still need to investigate the behaviour of a chained (^) operator.
It took me a long time to get to the current semantics.  Alas, I didn't
document properly at the time, so I will need to revisit that.  For the
record, the semantics of a chained (^) or definitely not the same for an
infix (^), especially when Bags/Mixes are involved.