Rewrites user example compilation test
Now everything is compiled via -c, instead of using EVAL. This is
slower, but closes #2764 since apparently the evaluation of so much
code was gobbling up memory, and closes #2782 since either the memory
itself or the creation of multiple objects was affecting this
somehow. It also causes #2789, so some changes are trying to fix that.