Using low-level parse to speed up things
Thanks to @ugexe and @taboege for hints on how to make this. I've
checked memory, and it's actually taking quite a bit of memory, but
not as much as EVAL and it's actually faster than that.

It's also producing a false positive with `class
Metamodel::ParametricRoleHOW` which I don't know how to fix.

Refs #2764 but leaves it open until

* Memory usage has actually been addressed, and
* We manage to fix that false positive