Changing back examples compilation
This closes #2782. The main problem is that some roles were defined
with `{ ... }`. Role definition *in the docs* affected example
compilation if it happened after that. This closes #2764 too. This is
related to #2789 although it was not a code quality issue per se, but
a documentation quality issue that was hidden in the previous version
of examples-compilation and suddenly showed up here.