Created a "Type" independent routines
This shows up as "routines" because it's not a class or anything, but
does not make htmlify.p6 choke.

This closes #2070. Also closes #2578 since we will not be creating
that unless this proves down the line to be untenable.

Still, #2059 has to be addressed. We have copied routines here, but
not eliminated them there.