Elaboration of Failures
Explain use fatal in place under Failure rather than just in pragmas,
and explain the `with` handling idiom.

The idiom is useful in explaining why one might use topicalization
with `else`, so also include that.

Finally, fix the with/orwith/without heading; unlike the other control
structures, it's lacking code tagging in the header.