CONTRIBUTING.md: refine passive voice principle
There's nothing wrong with the passive voice used correctly. Uh,
"there's nothing wrong with my using the passive voice?" "There's
nothing wrong with our using the passive voice?" Uh, no. ;-)

Technical documentation frequently needs the passive voice because the
agent is the "we" sometimes used in textbooks or tutorials consisting
of the reader and the writer which is better unstated.

I've changed the guideline to specify that what is discouraged is the
passive voice when the active voice is equally good--namely, in cases
where you'd need to use the word "by" in the passive.