Remove legacy issue_template.md; pull_request.md
Remove the legacy issue template.

Also, just a note that since I don't think we'll be using the multiple
pull-request templates feature (since it requires nonstandard tools
support) and the "new" location for a single global hidden template is
exactly the same as the "old" one, we can just leave it in place as-is.