Standardizes Pod name
To Pod6, which is exactly the same as the extension. It's bound to change eventually, so it's better if we always use the same thing.

I have changed Perl 6 Pod to Pod6 everywhere. Still, the official name with pod <non-breaking whitespace>6 is going to be used somewhere.

Pod has also been changed to Pod6, to avoid confusion with Perl 5's Pod. Several versions of the name, with or without capital, were used, so this also standardizes that.

This closes #1923