Includes the example and explanation
Adding commentary line complicated the grammar a bit too much, and I wasn't clear what the message was anyway. I guess that for exemplifying the reuse of grammars and corresponding actions it's OK, as well as for showing off some kinda complicated grammars.

This closes #1759, if this is not what you were looking for, feel free to clarify and reopen.