Add more details of sprintf and its format and *@args
Addresses Raku doc GH issue #3394

The '$' symbol has a special meaning in the sprintf format string and
must be escaped in a double-quoted format string if intended to be
used as a literal character. That fact is very briefly described in an
offhand manner, and it may result in a hard-to-find error due to an
error message that, up to Rakudo release 2020.05, does not mention it.

This commit should help by warning of a possible interpolated '$'
symbol in the format string upon input errors when the reported '%'
directives in the format string exceed the number of elements in the
*@args array.

The commit also expands information on the index directive as well as
adding a little more textual structure with =head3 Pod directives.

In addition, some other cleanups are introduced.

File modified:

+ doc/doc/Type/independent-routines.pod6