Replace 'RESUME:' with 'OUTPUT:' for consistency
The latest convention for the docs is to show output in examples in
this manner:

   ...some Raku code...   # OUTPUT: <<...some example output...\n>>

This commit updates examples to be consistent with the "# OUTPUT: ".
part of the convention.

It also: (1) adds missing French quotes and newlines around output
examples that are missing them and (2) aligns output examples
vertically when feasible.

Files affected:

+ doc/Language/exceptions.pod6
+ doc/Language/operators.pod6
+ doc/Language/regexes.pod6
+ doc/Type/IO/Handle.pod6
+ doc/Type/List.pod6
+ doc/Type/Map.pod6
+ doc/Type/Mu.pod6
+ doc/Type/Parameter.pod6
+ doc/Type/Range.pod6
+ doc/Type/Str.pod6
+ doc/Type/Supply.pod6
+ doc/Type/independent-routines.pod6