Simplify
It turns out that adding actions within the grammar is not (yet)
discussed in this document, and some aspects of it are different from
using an action object, so recommending the latter will need to be
qualified. As for the sentence that followed, well, it's true that
.parse does what it says there, but it has nothing to do with actions
being used or not, and having this here seems to me to be at best a bit
confusing.