Fix the INTERACT disambiguation mode
The code that processes the INDEX didn't match the format of the index
anymore, and probably was broken since the great list refactor a few
years ago. It was also much more complicated than necessary, since the
current index contains only simple Type::X entries.

So for now, make a very simple implementation that displays a list of
choices and lets the user pick one by number (or unique substring).

If any additional features are added to INDEX in the future, then this
can be replaced.