Fix indexing of proceed/succeed
Previously, there were two separate headers for proceed and succeed,
and likely meant to be read as "header + header + content".
Alas, on content pages used to display search results, a single pair of "header + content"
is delivered, and it is actually read as "header + nothing + header + content".
To avoid an empty page for the first header, just merge those two into a single one.