Expand on the use of {*} with proto
The documentation was not as clear as I needed. I tried several times to understand it, but failed. 
I got it nearly right when I asked on perl6-users.  
Crucially '{*}' is a special token without whitespace - this needs to be noted.  
The extra example combines the proto declaration with methods that use it.